+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.3.0
+
 * Add `process` and `go` metrics.
 * Add `channels_balance_satoshis` metric